4 - clean repo

+++ src/test/java/com/solab/iso8583/parse/CustomBinCodecs.java
@@ -86,7 +86,7 @@ public class CustomBinCodecs {
         }
         //Test parsing
         tmpl = mfact.parseMessage(buf, 0);
-        Assert.assertEquals(1234567890l, tmpl.getObjectValue(2));
+        Assert.assertEquals(1234567890L, (long) tmpl.getObjectValue(2));
         Assert.assertEquals(b29, tmpl.getObjectValue(3));
     }
